% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_bio}
\alias{get_bio}
\title{Get biography data for artists}
\usage{
get_bio(id)
}
\arguments{
\item{id}{QID}
}
\value{
tibble
}
\description{
Get biography data for artists
}
\examples{
get_bio(id = get_qid())
}
