% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{get_last_month}
\alias{get_last_month}
\title{Get first day of last month for pageviews functions}
\usage{
get_last_month()
}
\value{
a character vector containing timestamps that can be used with pageviews package.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
get_last_month()
}
