% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_net_worth}
\alias{get_net_worth}
\title{Get net worth of artists}
\usage{
get_net_worth(id)
}
\arguments{
\item{id}{QID}
}
\value{
Tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
get_net_worth(id = get_qid())
}
