% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_nominations}
\alias{get_nominations}
\title{Get number of nominations per artist}
\usage{
get_nominations(id)
}
\arguments{
\item{id}{Wikidata QID of artist}
}
\value{
Tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
get_nominations(id = get_qid())
}
