% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{get_pageviews_raw}
\alias{get_pageviews_raw}
\title{Get Wikipedia pageviews}
\usage{
get_pageviews_raw(
  project = "en.wikipedia",
  article = c("Taylor Swift", "Kanye West"),
  platform = "all",
  user_type = "all",
  start = get_last_month(),
  end = NULL,
  reformat = TRUE,
  granularity = "monthly"
)
}
\arguments{
\item{project}{the name of the project, structured as \code{[language_code].[project]}
(see the default).}

\item{article}{the article(s) you want to retrieve data for. Ideally features underscores in the title
instead of spaces, but happily converts if you forget to do this.}

\item{platform}{The platform the pageviews came from; One or more of "all", "desktop", "mobile-web" and
"mobile-app". Set to "all" by default.}

\item{user_type}{the type of users. One or more of "all", "user", "spider" or "automated". "all" by default.}

\item{start}{the start \code{YYYYMMDDHH} of the range you want to cover. This can be
easily grabbed from R date/time objects using \code{\link[pageviews]{pageview_timestamps}}.}

\item{end}{the end \code{YYYYMMDDHH} of the range you want to cover. NULL by default, meaning
that it returns 1 day of data.}

\item{reformat}{Whether to reformat the results as a \code{\link{data.frame}} or not. TRUE by default.}

\item{granularity}{the granularity of data to return; "daily" or "monthly", depending on
whether pageview data should reflect trends in days or months.}
}
\value{
data.frame
}
\description{
Get Wikipedia pageviews
}
\examples{
data <- get_pageviews_raw()
}
