% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_qid}
\alias{get_qid}
\title{Get QID for artists}
\usage{
get_qid()
}
\value{
Character vector of QID
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
For internal purposes.
}
\examples{
get_qid()
}
