% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{get_stats}
\alias{get_stats}
\title{Get statistics for artists}
\usage{
get_stats(id, language = "en")
}
\arguments{
\item{id}{QID}

\item{language}{Wikipedia project language}
}
\value{
Tibble.
}
\description{
Get statistics for artists
}
\examples{
get_stats(id = get_qid())
}
