% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{get_swift_spotify}
\alias{get_swift_spotify}
\title{Get Spotify data for Taylor Swift}
\usage{
get_swift_spotify()
}
\value{
List object.
}
\description{
Get Spotify data for Taylor Swift
}
\examples{
data <- get_swift_spotify()
}
