% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{visualize_pageviews}
\alias{visualize_pageviews}
\title{Visualize Wikipedia pageviews for Taylor Swift and Kanye West}
\usage{
visualize_pageviews()
}
\value{
ggplot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
visualize_pageviews()
}
