% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{get_spotify}
\alias{get_spotify}
\title{Get Spotify data for Taylor Swift and Kanye West in tidy format}
\usage{
get_spotify()
}
\value{
tibble
}
\description{
Get Spotify data for Taylor Swift and Kanye West in tidy format
}
\examples{
data <- get_spotify()

}
\seealso{
Other spotify: 
\code{\link{get_swift_spotify}()}
}
\concept{spotify}
