% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{get_spotify}
\alias{get_spotify}
\title{Get Spotify data for Taylor Swift and Kanye West}
\usage{
get_spotify()
}
\value{
List object.

tibble
}
\description{
Get Spotify data for Taylor Swift and Kanye West
}
\examples{
data <- get_spotify()
}
