% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{get_swift_spotify}
\alias{get_swift_spotify}
\alias{get_west_spotify}
\title{Get Spotify data for Taylor Swift and Kanye West}
\usage{
get_swift_spotify()

get_west_spotify()
}
\value{
List object.
}
\description{
Functions intented for backend only.
}
\examples{
data <- get_swift_spotify()
data <- get_west_spotify()
}
