% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_spotify}
\alias{theme_spotify}
\title{Apply spotify theme to a \code{gt} table}
\usage{
theme_spotify(gt_object, ...)
}
\arguments{
\item{gt_object}{gt table object}

\item{...}{Optional arguments}
}
\value{
gt table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Based on \code{gtExtras::gt_theme_dark()}.
}
