% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{visualize_spotify}
\alias{visualize_spotify}
\title{Visualize Spotify data for Taylor Swift and Kanye West}
\usage{
visualize_spotify()
}
\value{
gt table
}
\description{
Visualize Spotify data for Taylor Swift and Kanye West
}
\examples{
visualize_spotify()
}
