% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify.R
\name{visualize_spotify}
\alias{visualize_spotify}
\title{Visualize Spotify data for Taylor Swift and Kanye West}
\usage{
visualize_spotify()
}
\value{
gt table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
visualize_spotify()
}
